+++ src/app/new/components/mentionPopupItem.tsx
@@ -2,8 +2,8 @@ import * as Avatar from '@radix-ui/react-avatar';
 import { minidenticon } from 'minidenticons';
 import { useMemo } from 'react';
 
+import { displayNpub } from '@utils/formater';
 import { useProfile } from '@utils/hooks/useProfile';
-import { displayNpub } from '@utils/shortenKey';
 
 export function MentionPopupItem({ pubkey, embed }: { pubkey: string; embed?: string }) {
   const { isLoading, user } = useProfile(pubkey, embed);
