+++ src/app/settings/components/contactCard.tsx
@@ -5,7 +5,7 @@ import { useArk } from '@libs/ark';
 
 import { EditIcon, LoaderIcon } from '@shared/icons';
 
-import { compactNumber } from '@utils/number';
+import { compactNumber } from '@utils/formater';
 
 export function ContactCard() {
   const { ark } = useArk();
