+++ src/app/settings/components/zapCard.tsx
@@ -5,7 +5,7 @@ import { useArk } from '@libs/ark';
 
 import { LoaderIcon } from '@shared/icons';
 
-import { compactNumber } from '@utils/number';
+import { compactNumber } from '@utils/formater';
 
 export function ZapCard() {
   const { ark } = useArk();
