+++ src/shared/navigation.tsx
@@ -4,7 +4,7 @@ import { twMerge } from 'tailwind-merge';
 import { ActiveAccount } from '@shared/accounts/active';
 import { ChatsIcon, ComposeIcon, HomeIcon, NwcIcon, RelayIcon } from '@shared/icons';
 
-import { compactNumber } from '@utils/number';
+import { compactNumber } from '@utils/formater';
 
 export function Navigation() {
   const newMessages = 0;
