+++ src/shared/widgets/index.ts
@@ -15,3 +15,4 @@ export * from './other/toggleWidgetList';
 export * from './other/widgetList';
 export * from './other/addGroupFeeds';
 export * from './other/addHashtagFeeds';
+export * from './other/userProfile';
