--- src/utils/number.ts
@@ -1,2 +1,0 @@
-// convert number to K, M, B, T, etc.
-export const compactNumber = Intl.NumberFormat('en', { notation: 'compact' });
